+++ packages/_vue3-migration-test/src/main.ts
@@ -1,8 +1,9 @@
-import { XComponentsAdapter } from '@empathyco/x-types';
+import { QuerySuggestionsRequest, XComponentsAdapter } from '@empathyco/x-types';
 import { Component, configureCompat, createApp } from 'vue';
 import { createStore } from 'vuex';
 import { xPlugin } from '../../x-components/src/plugins/x-plugin';
 import { getRelatedTagsStub } from '../../x-components/src/__stubs__/related-tags-stubs.factory';
+import { getQuerySuggestionsStub } from '../../x-components/src/__stubs__/query-suggestions-stubs.factory';
 import App from './App.vue';
 import router from './router';
 import { facetsXModule, nextQueriesXModule, scrollXModule, searchXModule } from './';
@@ -31,6 +32,10 @@ const adapter = {
   relatedTags: () =>
     new Promise(resolve => {
       resolve({ relatedTags: getRelatedTagsStub(10) });
+    }),
+  querySuggestions: (request: QuerySuggestionsRequest) =>
+    new Promise(resolve => {
+      resolve({ suggestions: getQuerySuggestionsStub(request.query, 5) });
     })
 } as unknown as XComponentsAdapter;