+++ packages/_vue3-migration-test/src/x-modules/index.ts
@@ -7,3 +7,4 @@ export { default as TestElementsList } from './test-elements-list.vue';
 export * from './scroll';
 export * from './related-tags';
 export * from './history-queries';
+export * from './query-suggestions';