+++ packages/_vue3-migration-test/src/x-modules/query-suggestions/components/index.ts
@@ -1,0 +1,1 @@
+export { default as TestQuerySuggestions } from './test-query-suggestions.vue';