+++ packages/_vue3-migration-test/src/x-modules/query-suggestions/components/test-query-suggestions.vue
@@ -1,0 +1,9 @@
+<template>
+  <SearchInput />
+  <QuerySuggestions />
+</template>
+
+<script setup>
+  import SearchInput from '../../../../../x-components/src/x-modules/search-box/components/search-input.vue';
+  import QuerySuggestions from '../../../../../x-components/src/x-modules/query-suggestions/components/query-suggestions.vue';
+</script>